Convert all units into feet
This commit changes the code to use Feet when storing the values of the
curves. This is because feet/inches are the standard unit for FRC.

Now the canvas will convert the points on the canvas into feet using the
constant located in the image class (PIXELS_PER_FOOT). This is based on
the field image and must change if the image changes.